+++ src/app/center-bottom-content/components-bottom/report-builder/workflow/workflow.component.ts
@@ -2,7 +2,6 @@ import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 import { FormArray, FormBuilder, FormGroup } from '@angular/forms';
 import { Workflow } from 'src/app/shared/interfaces/workflow/workflow';
 import { MapService } from 'src/app/shared/services/map.service';
-import { NLDIService } from 'src/app/shared/services/nldi.service';
 import { WorkflowService } from 'src/app/shared/services/workflow.service';
 
 @Component({
@@ -20,10 +19,10 @@ export class WorkflowComponent implements OnInit {
   public stepsCompleted: number = 0;
   public numberOfSteps: number;
   public finalStep: boolean = false;
-  public delineationPolygon;
   public clickedPoint;
+  public selectedPerimeters;
 
-  constructor(private _fb: FormBuilder, private _nldiService: NLDIService, public _mapService: MapService, private _workflowService: WorkflowService) {
+  constructor(private _fb: FormBuilder, public _mapService: MapService, private _workflowService: WorkflowService) {
     this.workflowForm = this._fb.group({
       title: [],
       steps: this._fb.array([])
@@ -34,12 +33,12 @@ export class WorkflowComponent implements OnInit {
   ngOnInit(): void {
     this.setSteps();
 
-    this._nldiService.delineationPolygon.subscribe(data => {
-      this.delineationPolygon = data;
-    });
-
     this._mapService.clickPoint.subscribe((point: {}) => {
       this.clickedPoint = point;
+    })
+
+    this._mapService.selectedPerimeters.subscribe((perimeters) => {
+      this.selectedPerimeters = perimeters;
     })
   }
 
@@ -92,10 +91,20 @@ export class WorkflowComponent implements OnInit {
   }
 
   public subscription(i) {
-    if (this.workflowForm.value.title == "Delineation") {
-      this.workflowForm.value.steps[i].clickPoint = this.clickedPoint;      
-      this.workflowForm.value.steps[i].polygon = 'polygon';   
-    } 
+    switch (this.workflowForm.value.title) {
+      case "Delineation":
+        this.workflowForm.value.steps[i].clickPoint = this.clickedPoint;      
+        this.workflowForm.value.steps[i].polygon = 'polygon';   
+        break;
+      case "Fire Hydrology - Query Basin":
+        this.workflowForm.value.steps[i].clickPoint = this.clickedPoint;
+        this.workflowForm.value.steps[i].polygon = 'polygon';   
+        break;
+      case "Fire Hydrology - Query Fire Perimeters":
+        this.workflowForm.value.steps[i].clickPoint = this.clickedPoint;
+        this.workflowForm.value.steps[i].selectedPerimeters = this.selectedPerimeters;   
+        break;
+    }
   }
 
   public nextStep(step, value) {