+++ src/app/map/map.component.ts
@@ -4,13 +4,14 @@ import * as L from 'leaflet';
 import { Config } from 'protractor';
 import { ConfigService } from '../shared/config/config.service';
 import { HttpClient } from '@angular/common/http';
-import { NLDIService } from '../shared/services/nldi.service';
 import { WorkflowService } from '../shared/services/workflow.service';
 import "leaflet/dist/images/marker-shadow.png";
 import { ToastrService, IndividualConfig } from 'ngx-toastr';
 import * as messageType from '../shared/messageType';
 import * as esri from 'esri-leaflet';
 import { Workflow } from '../shared/interfaces/workflow/workflow';
+import { LoaderService } from '../shared/services/loader.service';
+import { AppService } from '../shared/services/app.service';
 
 @Component({
   selector: 'app-map',
@@ -18,28 +19,29 @@ import { Workflow } from '../shared/interfaces/workflow/workflow';
   styleUrls: ['./map.component.scss']
 })
 export class MapComponent implements OnInit {
-	public baselayers = [] as any;
   private configSettings: Config;
   private messager: ToastrService;
   public clickPoint;
   public currentZoom: number = 4;
   public latestDischarge: any;
-	public overlays = [] as any;
+	public workflowLayers = [] as any;
+	public activeWorkflowLayers:any;
   public selectedFeature: any;
   public marker: L.Marker;
   public basin: any;
   public splitCatchmentLayer: any;
-  public NHDLayer;
+  public traceLayer;
   public fitBounds: L.LatLngBounds;
   public selectedWorkflow: Workflow;
-  public delineationLoader: boolean = false;
+  public loader: boolean;
   public selectedPopup: any;
   public selectedSite: any
   public streamgageLayer: any;
+  public streamgageLayerStatus: boolean;
   public workflowData: any;
-  
+  public count: number = 0;
   constructor(public _mapService: MapService, private _configService: ConfigService, private _http:
-     HttpClient, private _nldiService: NLDIService, private _workflowService: WorkflowService, public toastr: ToastrService) { 
+     HttpClient, private _workflowService: WorkflowService, public toastr: ToastrService, private _loaderService: LoaderService, private _appService: AppService) { 
     this.configSettings = this._configService.getConfiguration();
     this.messager = toastr;
   }
@@ -57,6 +59,20 @@ export class MapComponent implements OnInit {
 
     // Add basemap
     this._mapService.SetBaselayer(this._mapService.chosenBaseLayerName);
+
+    // Get available workflow layers / Get active workflow layers
+    this.workflowLayers = this._mapService.workflowLayers;
+    this._mapService.activeWorkflowLayers.subscribe(activeLayers => {
+      this.activeWorkflowLayers = activeLayers;
+    })
+
+    // Get streamgages toggle status
+    this._mapService.streamgageLayerStatus.subscribe((status: boolean) => {
+      this.streamgageLayerStatus = status;
+      if (this.streamgageLayerStatus) {
+        this.setBbox();
+      }
+    })
 
     // Add scale bar
     this._mapService.scale.addTo(this._mapService.map);
@@ -91,20 +107,28 @@ export class MapComponent implements OnInit {
     this._mapService.clickPoint.subscribe((point: {}) => {
       this.clickPoint = point;
     })
-    // On map click, set click point value, for delineation
+    // On map click, set click point value
     this._mapService.map.on("click", (evt: { latlng: { lat: number; lng: number; }; }) => {
       this._mapService.setClickPoint(evt.latlng);
-      if (this.workflowData) {
-        if (this.workflowData.title == "Delineation" && this.workflowData.steps[0].completed) { 
-          this.onMouseClickDelineation();
+      if (this.selectedWorkflow) {
+        if (this.workflowData) {
+          if (this.workflowData.title == "Delineation" && this.workflowData.steps[0].completed) { 
+            this.onMouseClickDelineation();
+          }
+        }
+        if (this.selectedWorkflow.title == "Fire Hydrology - Query Basin") { 
+          this.onMouseClickFireHydroQueryBasin();
+        }
+        if (this.selectedWorkflow.title == "Fire Hydrology - Query Fire Perimeters") { 
+          this.onMouseClickFireHydroQueryFirePerimeter();
         }
       }
     }) 
     // On map zoom, set current zoom, display gages
     this._mapService.map.on('zoomend',(evt) => {
       this.currentZoom = evt.target._zoom;
+      this._mapService.setCurrentZoomLevel(evt.target._zoom);
       this.setBbox();
-      this.checkAvaliableLayers();
     }) 
     // On map drag, display gages
     this._mapService.map.on('dragend',() => {
@@ -113,45 +137,128 @@ export class MapComponent implements OnInit {
     // Subscribe to workflow
     this._workflowService.selectedWorkflow.subscribe((res) => {
       this.selectedWorkflow = res;
+      if (this.selectedWorkflow) {
+        if(this.selectedWorkflow.title != "Delineation") {
+          this.checkAvailableLayers();
+        }
+      }
+      if (!this.selectedWorkflow){
+        this.removeWorkFlowLayers();
+      }
     });
 
     //Subscribe to the form data
     this._workflowService.formData.subscribe(data => {
       this.workflowData = data;
-      this.checkAvaliableLayers();
+      if (this.workflowData) {
+        if (this.workflowData.title == "Delineation") {
+          this.checkAvailableLayers();
+        }
+      }
       if (!this.workflowData) {
-        this.removeLayer(this.NHDLayer);
+        this.removeWorkFlowLayers();
+      } 
+    });
+
+    //Subscribe to loader state
+    this._loaderService.loaderState.subscribe((state: boolean) => {
+      this.loader = state;
+    });
+
+    this.loadLayers();
+  }
+
+  public removeWorkFlowLayers(){
+    Object.keys(this.workflowLayers).forEach(layerName => {
+      this._appService.setLayerVisibility(layerName);
+      this._mapService.removeWorkflowLayers(layerName);
+      this.removeLayer(this.workflowLayers[layerName]); // No workflow is selected; remove all workflow overlayers
+    });
+  }
+
+  public checkAvailableLayers(){
+    if (this.selectedWorkflow) {
+      switch (this.selectedWorkflow.title) {
+        case "Delineation":
+          if (!this.activeWorkflowLayers.length) {
+            if (this.workflowData && this.workflowData.steps || this.activeWorkflowLayers.name != "NHD Flowlines") {
+              this.workflowData.steps[0].options.forEach(o => {
+                if (o.text == "NLDI Delineation" && o.selected == true) {
+                  this.addLayers('NHD Flowlines');
+                }
+              });
+            }
+          }
+          break;
+        case "Fire Hydrology - Query Basin":
+          this.addLayers('NHD Flowlines');
+          this.addLayers('Archived WildFire Perimeters');
+          this.addLayers('Active WildFire Perimeters');
+          this.addLayers('MTBS Fire Boundaries');
+          this.addLayers('Burn Severity');
+          break;
+        case "Fire Hydrology - Query Fire Perimeters":
+          this.addLayers('Archived WildFire Perimeters');
+          this.addLayers('Active WildFire Perimeters');
+          this.addLayers('MTBS Fire Boundaries');
+          this.addLayers('Burn Severity');
+          break;
+    }
+    }
+
+  }
+
+  private loadLayers() {
+    this.configSettings.workflowLayers.forEach(ml => {
+      try {
+        let options;
+        switch (ml.type) {
+          case "agsDynamic":
+            options = ml.layerOptions;
+            options.url = ml.url;
+            this.workflowLayers[ml.name] = esri.dynamicMapLayer(options);
+            break;
+          case "agsFeature":
+            options = ml.layerOptions;
+            options.url = ml.url;
+            this.workflowLayers[ml.name] = esri.featureLayer(options);
+            if (this.configSettings.symbols[ml.name]) { 
+              this.workflowLayers[ml.name].setStyle(this.configSettings.symbols[ml.name]); 
+            }
+            break;
+        }
+      } catch (error) {
+        this.createMessage(ml.name + ' layer failed to load','error');
       }
     });
   }
 
-  public checkAvaliableLayers(){
-    if (this.currentZoom >= 8 && this.workflowData && this.workflowData.title === 'Delineation') { // checking current zoom and workflow
-      this.workflowData.steps[0].options.forEach(o => {
-        if (o.text == "NLDI Delineation" && o.selected == true) {
-          this.addLayers('NHD');
-        }
-      })
-    }
-  }
-
   public addLayers(layerName: string) {
-    if (layerName == "NHD") {    //Setting stream layer
-      this.NHDLayer = esri.dynamicMapLayer({
-        'url': 'https://hydro.nationalmap.gov/arcgis/rest/services/nhd/MapServer',
-        'layers': [5,6]
-      })
-      this.NHDLayer.addTo(this._mapService.map);
-    }   
+    this.configSettings.workflowLayers.forEach((layer: any) => {
+      if (layer.name === layerName) {
+        layer.visible = true;
+        this._mapService.setWorkflowLayers(layer);
+      }
+    }); 
+    this.workflowLayers[layerName].addTo(this._mapService.map);
   }
 
   public setBbox(){
-    if (this.streamgageLayer !== undefined) this._mapService.map.removeLayer(this.streamgageLayer);
+    if (this.streamgageLayer !== undefined) {
+      this._mapService.map.removeLayer(this.streamgageLayer);
+      this.configSettings.overlays.forEach((overlay: any) => {
+        if (overlay.name === "Streamgages") {
+            overlay.visible = false;
+        }
+      });
+    }
     if (this.currentZoom >= 8) {
       var bBox = this._mapService.map.getBounds();
       var ne = bBox.getNorthEast(); // LatLng of the north-east corner
       var sw = bBox.getSouthWest(); // LatLng of the south-west corder
-      this._mapService.setStreamgages(sw.lng, ne.lng, sw.lat, ne.lat)
+      if (this.streamgageLayerStatus) {
+        this._mapService.setStreamgages(sw.lng, ne.lng, sw.lat, ne.lat)
+      }
     }
   }
   
@@ -182,6 +289,13 @@ export class MapComponent implements OnInit {
           return L.marker(latlng,{icon: MyIcon});
         }
       }).addTo(this._mapService.map);
+      this.configSettings.overlays.forEach((overlay: any) => {
+        if (overlay.name === "Streamgages") {
+          overlay.visible = true;
+        }
+      });
+      // setting streamgages layer in map services
+      this._mapService.setStreamgagesLayer(this.streamgageLayer)
     }
   }
 
@@ -202,10 +316,10 @@ export class MapComponent implements OnInit {
     this.removeLayer(this.splitCatchmentLayer);
     this.addPoint(this.clickPoint);
     this.marker.openPopup();
-    this.delineationLoader = true;
-    this.createMessage("Delineating Basin. Please wait.");
-    this._nldiService.getUpstream(this.clickPoint.lat, this.clickPoint.lng, "True");
-    this._nldiService.delineationPolygon.subscribe((poly: any) => {
+    this._loaderService.showFullPageLoad();
+    this.createMessage("Delineating basin. Please wait.");
+    this._mapService.getUpstream(this.clickPoint.lat, this.clickPoint.lng, "True");
+    this._mapService.delineationPolygon.subscribe((poly: any) => {
       this.basin = poly.outputs;
       if (this.basin) {  
         this.removeLayer(this.splitCatchmentLayer);  
@@ -213,8 +327,99 @@ export class MapComponent implements OnInit {
         this.splitCatchmentLayer.addTo(this._mapService.map);
         this._mapService.map.fitBounds(this.splitCatchmentLayer.getBounds(), { padding: [75,75] });
       }
-      this.delineationLoader = false;
+      this._loaderService.hideFullPageLoad();
     });
+  }
+
+  public onMouseClickFireHydroQueryBasin() { 
+    // Issue #57: see onMouseClickDelineation() to start
+  }
+
+  public onMouseClickFireHydroQueryFirePerimeter() { 
+    this._loaderService.showFullPageLoad();
+    this.count = 0;
+    this.createMessage('Querying layers. Please wait.');
+    Object.keys(this.workflowLayers).forEach(layerName => {
+      if (layerName === 'Active WildFire Perimeters' || layerName === 'Archived WildFire Perimeters') {
+        this.workflowLayers[layerName].query().nearby(this.clickPoint, 4).returnGeometry(true)
+          .run((error: any, results: any) => {
+            this.findFeatures(error,results,layerName);
+          }
+        );
+      } else if (layerName === 'MTBS Fire Boundaries') {
+        this.workflowLayers[layerName].identify().on(this._mapService.map).at(this.clickPoint).returnGeometry(true).tolerance(5)
+          .run(async (error: any, results: any) => {
+            this.findFeatures(error,results,layerName);
+          }
+        );
+      }
+    });
+  }
+
+  public async findFeatures(error,results,layerName) {
+    let popupcontent;
+    let selectedPerimeters = [];
+    let layer;
+    const shownFields = ['INCIDENTNAME', 'COMMENTS', 'GISACRES', 'FIRE_YEAR', 'CREATEDATE', 'ACRES', 'AGENCY', 'SOURCE', 'INCIDENT', 'FIRE_ID', 'FIRE_NAME', 'YEAR', 'STARTMONTH', 'STARTDAY', 'FIRE_TYPE'];
+    if (error) {
+      this.createMessage('Error occurred.','error');
+      this._loaderService.hideFullPageLoad();
+    } 
+    if (results && results.features.length > 0) {
+      results.features.forEach(feat => {
+        popupcontent = '<div class="popup-header"><b>' + layerName + ':</b></div><br>';
+        if (layerName === 'MTBS Fire Boundaries') {
+          let date = feat.properties.STARTMONTH + '/' + feat.properties.STARTDAY + '/' + feat.properties.YEAR;
+          if (date.indexOf('undefined') > -1) date = 'N/A';
+        }
+        Object.keys(feat.properties).forEach(prop => {
+          if (shownFields.indexOf(prop.toUpperCase()) > -1) {
+            let val = feat.properties[prop];
+            if (prop.toLowerCase().indexOf('date') > -1) {
+              val = new Date(val).toLocaleDateString();
+            }
+            popupcontent += '<b>' + prop + ':</b> ' + val + '<br>';
+          }
+        });
+        popupcontent += '<br>';
+        if (layerName === 'MTBS Fire Boundaries') {
+          layer = L.geoJSON(feat.geometry);
+        } else if (layerName === 'Active WildFire Perimeters' || layerName === 'Archived WildFire Perimeters') {
+          const col = layerName.indexOf('Active') > -1 ? 'yellow' : 'red';
+          layer = L.geoJSON(feat.geometry, {style: {color: col}});
+        }
+        layer.addTo(this._mapService.map);
+        this.addBurnPoint(layer.getBounds().getCenter(), popupcontent);
+      });
+      selectedPerimeters.push({ 'Key': layerName, 'Data': results})
+      const data = await this._mapService.trace(results).toPromise();
+      this.addTraceLayer(data);
+    }
+    this._mapService.setSelectedPerimeters(selectedPerimeters);
+    this.count ++;
+    this.checkCount(this.count, 3);
+  }
+
+  public addTraceLayer(data) {
+    this.traceLayer = L.geoJSON(data);
+    this.traceLayer.addTo(this._mapService.map);
+    this._mapService.map.fitBounds(this.traceLayer.getBounds(), { padding: [75,75] });
+    this._loaderService.hideFullPageLoad();
+  }
+
+  public checkCount(count, goal) {
+    if (count === goal) {
+      if (this.loader == true) {
+        this._loaderService.hideFullPageLoad();
+        this.createMessage('Must select a fire perimeter.','error');
+      }
+    }
+  }
+
+  public addBurnPoint(latlng, popupcontent) {
+    this.marker = L.marker(latlng).bindPopup(popupcontent).openPopup();
+    this._mapService.map?.addLayer(this.marker);
+    this.marker.openPopup();
   }
 
   public addPoint(latlng: any) {
@@ -232,7 +437,7 @@ export class MapComponent implements OnInit {
 
   private createMessage(msg: string, mType: string = messageType.INFO, title?: string, timeout?: number) {
     try {
-      let options: Partial<IndividualConfig> = undefined;
+      let options: Partial<IndividualConfig> = null;
       if (timeout) { options = { timeOut: timeout }; }
       this.messager.show(msg, title, options, mType);
     } catch (e) {