+++ src/app/shared/interfaces/config/config.ts
@@ -7,4 +7,6 @@ export interface Config {
     GageStatsServices: string;
     GageStatsServicesBounds: string;
     baseLayers: [];
+    workflowLayers: [];
+    overlays: [];
 }