+++ src/app/top-bar/top-bar.component.spec.ts
@@ -1,5 +1,4 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { TopBarComponent } from './top-bar.component';
 
 describe('TopBarComponent', () => {