+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Added general map functions: scale, zoom, show your location, geosearch, basemap selection
 -   Added google analytics
 -   Added basin delineation workflow
+-   Added fire hydrology workflow
+-   Added trace functionality for query fire perimeters
+-   Added toggleable layers
 
 
 ### Changed